refactor(gitlab_api): Remove references to `generate_gitlab_full_configuration`
This function was used in very similar contexts and for similar purposes as `get_gitlab_ci_configuration`, we try to consolidate both into a single function
Note that the new function is slower than the original, but generates a more "complete" gitlab config - the old config objects' `.keys()` were subsets of the new objects'.
I will change the usages of this new function, adding flags to match the speed of the original.