refactor(tasks): Rename `with_lint` argument to `partial_resolve`
This argument was named this way because it controls whether `resolve_gitlab_ci_configuration` calls the `/lint` endpoint - but that endpoint actually does more important things than just lint the config file. It is actually mostly used for resolving `extend` and `!reference` tags in the configs, which is almost always what we want.