Search image based on its name or repo digests
As stated in the comments for container image:
For historical reason, The imageId from containerd runtime and kubernetes refer to different fields.
For containerd, it is the digest of the image config.
For kubernetes, it referres to repo digest of the image (at least before CRI-O v1.28)

So we look for the image whose name is either the container image ID
or it has it in its repo digests