service discovery: Rework network stats interface
Rework the network stats interface to a single call which gets the list
of PIDs to be monitored.  This will allow the future split of the
handling between the core agent and system-probe to be simpler, avoiding
multiple remote calls and removing the need for special handling to
synchronizing state in case either side restarts.