[RC] Prevent error logs on client requests when RC is uninitialized
In a recent version of the agent, we added a new behavior that would
cause the Remote Config service running in the agent to send a "config
flush" message to clients when the TUF metadata had expired. This check
requires being able to pull the Director Repo TUF metadata, but this
repo is empty if Remote Config is uninitialized (either RC is disabled
for the org, or the API key is not scoped for RC). This was causing an
error message "empty targets meta in director local store" to be sent to
clients, which users were seeing in their application logs and
rightfully being concerned about. Since this message is not an actual
error and is not actionable, it's not helpful and spammy for users.

We don't have any expectations that an uninitialized repo becomes
initialized over time, so we don't need to have a deadline check in this
block of code. If the repo is uninitialized, there are no config files
in clients, and thus no need to ever send a flush cache message.