usm: tests: Use t.Cleanup instead of local defer to cleanup unused resources
We used 'defer cleaners()' for the cleaners callback returned from GetProtocolStats.
However, we stored locally entries from the protocol-stats, but it was already cleaned.
That can lead into flakes in the CI, as we modify sketches that were already returned to
the shared pool. Thus, we couldn't guarantee a clean ddsketch

The use of 't.Cleanup(cleaners)' ensures we don't return to the pool any ddsketch object
until the test is over.