[DEBUG-3986] dyninst/actuator: add a stateful control layer
This layer orchestrates the lifecycle of programs and how
they map to processes. Most of the implementation and testing
goes into the state machine and its updates. I suspect it could
have a better implementation -- right now it's just a bunch of
imperative case crushing. I tried to make up for that with
a good deal of testing and testing infrastructure.

The testing comes in two main forms:
 - TestSnapshot: uses golden files in a yaml DSL to list the
   set of events that should happen and then
   reports the state transitions and effects
 - TestStateMachineProperties: uses random event generation
   to try to shake out interesting behaviors

Both of these tests rely on invariant checking of the state
itself.

In addition to the state, there's the implementation of the
Actuator which hooks up the state machine and the effects it
generates to real dependencies.

One thing this change makes abundantly clear is the need to
have fine-grained error reporting out of ir generation, and
then separately a clearer interface for how to report these
errors out into the diagnostics flows.