[usm] fix generation script for debug info lookup table
Before this patch, the generation of the gotls/lookup/luts.go file was
failing: Go versions below 1.16 cannot parse the go.mod file of the
x/net module at versions above 0.35 (released a few months ago). This
patch pins the module at v0.35. We were arbitrarily using the latest
x/net version (and thus using the debug info for the latest); now we're
arbitrarily using a fixed version. Luckily, the debug info we need for
x/net is stable -- we need the offset of an embedded field in a struct
(*).

In addition to failing on Go 1.16, the latest version of x/net was
forcing the toolchain selection mechanism in go 1.21+ to select 1.23 --
so we were not actually using go 1.21 and 1.22. The next commit makes
the script robust to this.

The regenerated luts.go doesn't have any changes. FWIW, I think this shows
that the debug info in that file is valid for Go 1.24; I'm not sure we
ever ran the script for 1.24, since I think the release of x/net that
broke it happened a little before 1.24.