[usm] inhibit Go toolchain selection when generating luts.go
The generator script for luts.go intends to use every Go compiler
version. Until this patch, it was fragile because it allowed Go's
toolchain selection mechanism to transparently use a newer toolchain
version than the one the script was intended to use, subject to the
requirements of the modules used by the test program. This could lead to
the generated source containing debug info for the wrong compiler
versions. In fact, this was happening until the previous commit.

This patch makes the generation robust by forcing the intended toolchain
versions.