Fix overlayfs based SBOM generation method
PR #35389 bumped our Trivy fork causing the generation
of SBOM using overlayfs to fail.

The method modifies the passed options to the walker,
appending the path of the image layer which causes all
the folders to be skipped because of OnlyDirs.

Long term solution would be to not modify the walker
options. In the meantime, we remove the prefix added
by container.NewArtifact.