fix(linter): Use a minimally-resolved config object in `gitlab_ci`
This is done for performance reasons, and is the way it is currently done on `main`, but I accidentally changed during the refactor.
Without this change the `dda inv gitlab-ci -t main` invocation takes ~60s, up from ~40s as it is currently on `main`. This is due to the extra time taken by creating a fully-resolved gitlabci config object instead of the actually-needed minimal one.
This commit makes sure we use a minimally-resolved one, restoring the performance currently on main.