fix(linter): Remove postprocessing clean and filtering from `test_gitlab_configuration`
The config object passed to `test_gitlab_configuration` is minimally-resolved (only `include`s are resolved), for performance reasons.
Before testing the validity of the config object, we should not be removing any dotted jobs, `extends` tags or the like, otherwise the configuration will be considered invalid.
Thus we need to disable the `clean` and `do_filtering` options from the post-processing function, as that is exactly what they do.