(fleet) fix integrations restore python script file permissions on linux
This PR fixes a permission bug in the python script used to restore integrations.

The scripts were lacking any enforced mode when creating their `.txt` files. This defaulted to 0666 for files + the system-wide umask. With the migration to the installer the python script are executed in a context where umask is 0, exposing this bug even more.

To fix this I enforced a chmod 0644 for all the files the script creates.