bininspect: switch pointer to value receiver
GoArch is a string type that had a pointer receiver method. I don't see
any point to the pointer, and it having a pointer receiver precludes the
method from being called on temporaries. We have code returning GoArch by
value (as is natural), so the pointer receiver forced you to give a name
to such return values. This patch switches it to value receiver.