dyninst: fix frameless function CFA computation
This patch does two things:
1) It propagates the frameless nature of functions.
2) It appropriately initializes the CFA based on that frameless nature.

Go leaf functions that don't use stack space do not have a prologue. We
handled this computation for deciding where to put the probe, but we
didn't translate the CFA offset calculation. Note that this fix remains
incomplete in that it's possible for inlined functions to appear underneath
a frameless parent. Plumbing for that calculation is left for later.