Support for AWS SecretsManager in Secrets Helper
It's possible that the datadog agent is deployed in containerized
environments in AWS where it isn't feasible to provide a custom secrets
helper executable to fetch secrets from AWS Secrets Manager. AWS Secrets
Manager is a common way to provide secrets to containers in a secure
way.

This change extends the supported secrets providers to include AWS
Secrets Manager. By providing an ARN to a secret as the id, the secrets
helper can pull the secret from Secrets Manager provided that the
agent was granted read access to the secret.