dyninst/actuator: create Tenant concept
There are multiple users of the actuator that might each want to create
different programs on a per process basis and yet we want them all to
share a single set of resources.

This change feels right insofar as we do want to isolate components that
want to load programs and fit into the program lifecycle independently.
A weird property of this change is that for there to be multiple
programs applied to the same process, they *must* be by different
tenants. This is not different than before -- it is more relaxed than
before, but it still feels a bit off.