Promote uds_dogstatsd_to_api_cpu to quality_gate_metrics
This commit promotes the pre-existing `uds_dogstatsd_to_api_cpu` to a
new `quality_gate_metrics` experiment, focused on CPU and asserting
CPU load in a non-idle environment. It is understood as of this writing
that this experiment represents a high-load scenario but given the long-term
nature of this experiment in our CI without comment we understand this scenario
to have importance, hence its promotion.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>