[incident-40404] Pass `ddtrace`'s `crashtracker` to `codesign`
Notarization requests started failing silently seemingly due to an error
code being swallowed, so we didn't notice the binary was not
`codesign`ed.

Current macOS deliverables (as of Agent release 7.68.0) are therefore
affected, hence the need for a short-term countermeasure that will buy
us some time for a proper fix.