Add fuzz tests for docker stream parsing, fix truncation panic
This commit adds fuzz tests for the Docker stream parsing in pkg/logs. These
tests demonstrated a panic in the event of a truncated message on the stream,
owing to malformed Docker response (unlikely) or a severed connection. This
panic was re-confirmed with a unit test and corrected.

None known.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>