usm: kafka: tests: Enable kafka tests on ubuntu 24.10
We believed there was issue with kafka-tls on top of ubuntu 24.10, hence we disabled those tests
when we run on ubuntu 24.10. however, we discovered the issue was due to a breaking change in
our GoTLS monitoring on kernels 6.11 and later.

The original issue was fixed, so we can enable kafka tests again on ubuntu 24.10