Introduce fuzz tests for our kubernetes logs parser
This commit introduces a fuzz test for our kubernetes logs parsing. We
previously had a FIXME in the tests that indicated timestamp handling was not
strict and on review if k8s kicks an invalid timestamp to use we will then
potentially use that in the tailer mechanism as a restart time. This FIXME is
now removed and corrected.

The fuzz tests found no other odd behavior here.

Peer to #38579

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>