Increase -all-features memory limit to 610MiB
We increased the memory limit for this experiment in PR #38497. That limit
increase demonstrates median instance behavior based on the 2025/07/09 nightly
run of Quality Gates. We believe 610 MiB represents a comfortable overhead based
on that nightly and as we gather more understanding of Agent's new behavior in
this experiment we can reduce.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>