dyninst/rcscrape: do not load dwarf to scrape
This change implements a custom ir loader that will not use so much
memory or compute to find probes by not using dwarf information. 
Instead we rely on the fact that the ABI is stable and we know that
these functions we want to probe are also well-defined as frameless
functions with well-known arguments.