dyninst: add a more efficient way to resolve PCs
This patch adds a facility for resolving multiple increasing program
counters within a single function more efficiently than repeated calls
to GoSymbolTable.LocatePC(), by amortizing the lookups for the function
information. When exploring DWARF for SymDB we're generally in the
context of a function and we want to resolve many program counters
within it before moving on to the next function.

The new FuncPCResolver also has functionality for figuring out the line
number corresponding to the end of a function definition, useful for
SymDB.