dyninst/procmon: cache executable analysis
Analyzing an executable is not free. For a big binary it can be a few
ms. There are cases where executables fork children frequently or get
caught in loops and to incur milliseconds of CPU time per process launch
could be rather painful. This PR adds caching to make the checking
very cheap in the case that we know the binary is not interesting.