[APM] Increase TestStopWaits test timeout
This test can be flaky if the test agent has not started its goroutines,
causing the timeout condition to trigger.

A simple fix is to raise the timeout to 500ms. If test remains flaky
with this setting, we should consider a different approach, perhaps by
checking the readiness of the agent workers.