fix(e2e): Don't overselect tests on first run of e2e tests
Instead of trying to select *exactly* the tests passed in argument to `new-e2e-tests.run`, restore the old behavior of passing the argument more-or-less directly to `go test`. This way we ensure all existing usages of `--run` in CI keep their old behavior, and the very specific selection of tests to run is kept only when attempting to retry tests.