Update macOS installation script for AArch64/ARM64
The script doesn't need to wait for AArch64/ARM64 DMGs to be publicly
released - quite the other way around as we're going to add an
architecture qualifier to DMGs, see: DataDog/omnibus-ruby#240.

The present change therefore adapts the script to the foreseeable
future, by considering the optional presence of an architecture
qualifier.