feat(rc): implement WebSocket echo test
This commit implements a simple WebSocket echo test, connecting to the
RC backend (at "/api/v0.2/echo-test") and exchanging a series of data
frames.

This test allows RC to evaluate the connectivity and performance of
WebSockets in the wild, as a potential transport for a new protocol
between the Agent and RC.

Part of RC-2602.