dyninst/object: handle binaries with mixed dwarf
We ran into an issue in the wild where a binary had some compile units
at dwarf 5 from a C compiler and the go was using dwarf 4. Our logic to
decide which section to use to read locations lists was too brittle --
it just preferred the loclists section if it existed. This fixes that
bug and adds testign that would have caught it.