dyninst/irgen: refactor for efficiency
This started primarily as a frustration that irgen was slow. It turned
out to be slow because of the quadratic use of the line reader. This
is fixed in this PR by deferring when we go and search for prologues and
then doing it in sorted order. Along the way I ended up refactoring the
main logic of generateIR to be more decomposed. I think it's better like
this.

Before: `TestIRGenAllProbes/sample (20.02s)`
After: `TestIRGenAllProbes/sample (2.82s)`