[RC] Adjust default core service poll interval
The agent has the default poll interval hardcoded to 60s, but we
override this to 50s for every agent where possible. This causes some
confusion if you're only reading the agent codebase and not aware of our
backend, so to make this more clear adjust the default interval in the
agent.