dyninst/symdb: support filtering out 3rd-party symbols
For purposes of saving on both resources while processing DWARF and the
amount of data produced for SymDB, we'll want to restrict the symbols we
collect to 1st party code (at least optionally). This patch teaches the
extractor to distinguish 1st party vs 3rd party. For binaries with
buildinfo (i.e. not built by Bazel), we look at the URL of the main
module and select the github org. For Bazel binaries, we exclude some
file prefixes.