dyninst/procmon: fix memory leak when nothing is interesting
Prior to this change, we'd always add processes to the pending map but we'd
only ever clear the map when we had something interesting to report. This
results in a memory leak in the common situation when no processes are interesting
to dynamic instrumentation.

Testing was expanded to assert that the pending map is always a subset of the alive
map and the code was updated to uphold this invariant.