Secret backend test uses prebuilt binary
Previously, this test would compile multiple executable binaries in order to test the secret backend command. This is both (1) inefficient and (2) flaky. This PR replaces the multiple executables with a single binary while still covering the same functionality. Additionally, the test files contain prebuilt binaries for a couple of platforms that we support, and the test will use this if it matches the OS and CPU architecture, invoking the go compiler only if no prebuilt binary is found.