Also make `curl` retry in macOS install script
This aims at making the script slightly more resistant to transient
network errors, similarly to what's been done otherwise, for instance:
- `cmd/agent/install_script.sh`:
  https://github.com/DataDog/datadog-agent/blob/877f00e9765b2e6bd9f95a6238cd3438cf684799/cmd/agent/install_script.sh#L493
- `pkg/fleet/installer/setup/install.sh`:
  https://github.com/DataDog/datadog-agent/blob/877f00e9765b2e6bd9f95a6238cd3438cf684799/pkg/fleet/installer/setup/install.sh#L60

The proposed number of retries is here of `2`, which means at most `3`
attempts will be made. (instead of `1`)

Transient error means either: a timeout, an HTTP 408, 429, 500, 502, 503
or 504 response code.