Converge package quality gates. [Part 1]
- Introduce functions in new gates_lib that
are required to work with packages (deb, rpm, heroku).
Processing is also simplified. For instance, flavor
is no longer needed to determine path to the package.
Furthermore, kwargs are not used to make functions
stricter.

- Add necessary unit tests to test
newly introduced or converged functions.

The implementation is not done. Docker
and MSI are still not supported. Most probably
the initialization logic of the StaticQualityGate
class will be extended later on to accomodate Docker
and MSI.