feat(rc): set User-Agent header for RC requests
This causes all Agent HTTP (and WebSocket handshake) requests to the
remote config backend to present a user agent in the form of:

    datadog-agent/7.68.2 (go1.24.5)

While the Agent does declare its version in some request payloads, this
only works for those specific requests, and only if those payload can be
successfully parsed.

Setting the user agent on the HTTP request can be used to identify the
version of the agent making a request to RC without attempting to parse
the payload, prior to reading the body at all.