dyninst/ebpf: use correct length for variable sized types
Before this change, if we had a slice or something variable sized, we'd
assume that the number of bytes we'd read the right amount but then try to
enqueue on the scratch buffer the maximum number of bytes. This had a fun
side-effect of only surfacing issues in the test that runs all the probes
because in all the other tests, the scratch buf is zeroed. Elsewhere we
should assert on some other behaviorial statistics of probes like the number
of pointers we chased and we should also really mess with the initial
contents of the scratch buffer.