Add WMI database cleanup for windows docker
We have seen in the past that one particular
windows docker image's size varies in a range
of 25 Mb. Layer by layer analysis shows that
it is related to OBJECTS.DATA file that is
created and used by WMI as a database.
Sometimes cleanup kicks in before RUN
of Dockerfiles/agent/install.ps1 finishes,
sometimes it doesn't. That is the reason why
the size varies.
We are adding an unconditional WMI database
cleanup step. It is allowed to fail, because
non-core images do not have WMI, therefore,
script fails.