dyninst/ebpf: add a trie for pointer tracking
Before this change we just used a list. This meant that we
needed to keep the number of pointers we could chase to be
quite small because we didn't want quadratic behavior. Before
this change, the limit on the number of pointers was 128 in
total. Now that number has been lifted to 1024. The memory
usage has gone up to an extra 4 bytes per pointer: from
1536 total to 16384. This is still quite efficient given
what we're doing.

The structure can, with the same widths, be used for up to
2048 pointers which will be exactly 32KiB in size. This is
handily the maximum size for a single per-cpu array item.

It will take some more clever tricks to extend the structure
beyond 2048 items, but we can cross that bridge when we get
there.