dyninst/symdb: introduce a streaming api
Introduce an iterator over a binary's packages so that the symbols can
be streamed out without accumulating them all in memory. For now this
implies configuring the extraction process to ignore abstract functions
and inlined instances, since they can appear in compile units
corresponding to other packages.

Even when using this streaming API, there is still a types cache that
accumulates types across packages (populated when resolving variables).
We'll see if that's a problem for memory use.