dyninst/symdb: fix handling of types in packages with dots
Fix a bug where types from packages containing a dot in the component of
the import path where added to the wrong package, or sometimes not added
at all. We were using the unescaped import patch when parsing out the
type's package, which made the regex we use for parsing incorrect. Now
we use unescaping utilities copied from the compiler, which are more
robust.

Also apply unescaping to the type names of variables; before, we were
producing variables with escaped type names in the output.