dyninst/object: add DiskCache for decompressed sections
This commit introduces a DiskCache that can be used to load
`ElfFile` objects. The returned ElfFile will have its compressed
`DebugSections` data in mmaps backed by files managed by the
cache. The cache will track the references to the files and
will clean them up when they are no longer needed.

Additionally, the cache can be configured to enforce limits on
 * the amount of disk space that can be used in aggregate, 
 * the amount of disk space that must be left available on the
   device, and
 * the percentage of disk space that must be left available on
   the device.

The cache isn't really a "cache" per se in that it doesn't keep
the files around when they aren't in use, but that will be a
trivial extension.