dyninst/actuator: move elf loading to IRGenerator
This commit slightly reworks the dependencies of the
actuator such that now elf file loading is delegated to
the ir generator as opposed to happening unconditionally
in the actuator.

This was something of a major oversight in that in meant
that we were always decompressing the debug sections just
to do remote config scraping.