dyninst/symdb: don't ignore packages that only have types
Before this patch, we would not output information about types from
packages that have no functions (e.g. packages that only contain
exported structs).

Also this patch makes it so that we don't output empty types (types with
no fields or methods), on the argument that they wouldn't be useful to
either the backend or the UI.