dyninst/symdb: refactor the package iterator
Before this patch, extracting symdb symbols needed two calls -- one to
construct a "builder" (NewSymDBBuilder) and then another to actually get
some data (PackagesIterator(), or ExtractSymbols()). It was unclear if
the same builder can be used more than once, which made it unclear why
there are two steps to begin with. This patch clarifies the API by
exposing ExtractSymbols() and PackagesIterator() as standalone
functions, and hiding the iteration state in an unexported struct.