Global Datadog() can only read/write config, not build it
To more cleanly separate the building and usage of the config, remove all model.Setup methods from the interface that is returned by the lglobal Datadog() accessor.

Introduce the model.BuildableConfig interface, which is able to both build and read/write the config. It is only meant to be used during initial construction.

Finally, introduce config.NewBuilder() which is an escape hatch for OTel, since they build on top of the already built config.