usm,cnm: Ensure tag index is always positive
Agent without agent-payload v5.0.162 will provide the default number of the host tags index and local containers tag index
which is 0. However, 0 is a valid tags index, and can cause us to incorrectly add irrelevant tags or just waste CPU and memory
in the backend. Hence, we force the tagger to provide index > 0.