dyninst/symdb: teach CLI to upload symbols
This should be useful for manual testing, but also for dogfooding where
we don't have a new-enough agent or dd-trace-go.

The upload comes from either a local file or a container image. For the
latter, you can now give the tool the name of an image from
registry.ddbuild.io and it will look for a file with the same name
inside the image's /usr/local/bin. This should make it easy to upload
symbols for services running in production by clicking through the
Datadog UI to find these image names.