process_log: Do not use tracer service name
Currently, if a tracer is added to a service, the service name for
discovered process logs may change, even without linking logs and
traces.  This behaviour is different from how things work for discovered
container logs.  Change the process behaviour to match the container
behaviour:

Now, just adding a tracer will not change the service name.  If logs and
traces are linked, then the tracer's injected attributes will lead to
correct service names for the logs on the backend.  Also, if a
DD_SERVICE is manually set, then that name will be used in the logs
configuration.