Proof of concept of user customizable agent devcontainer setup.
We introduce the notion of $HOME/.devcontainer/agent_overrides.json.

If it exists, we load it and add to the defaults done by inv devcontainer.setup.

It's useful enough just to make it easy to permanently get some things
you want mounted always appear in the container.

The bit about appending the createCommand is dubious.

But there is so much more we could do. It feels like this command
should be external to the agent code because it is not bound to
a particular agent branch, it is more tied to the current container
image we build.   If we made it external you could make multiple
containers with different options, and stil work on the agent.
```
agent_devcontainer create -name dev_normal
agent_devcontainer create -name dev_with_foo --extramount $HOME/foo:/home/datadog/foo
agent_devcontainer create -name dev_cursor --extramount $HOME/my_prompts:/home/datadog/my_prompts
```

Since that is the goal, I don't like the fixed path to agent_overrides.json.
That should also be a command line parameter.  But I haven't figure out how to
add command line args to invoke tasks yet.