dyninst: collect interface data
This commit adds support for resolving the dynamic type underneath
interfaces in ebpf, and adds support for writing them out in the
decoder. It has some known missing pieces such as not resolving
types we did not compile into the program to names, and not adding
types that might make sense given interface values or well-known
primitives.