Cross-org telemetry for DogStatsD metrics type mixture
This commit introduces COATs metrics for the metrics type mix we get
in DogStatsD traffic. We understand that dogstatsd performance is
heavily dictated by the metrics mixture an instance of dogstatsd
receives. The goal of this change is to record this measure.

I've introduces new metrics in the server `metric_type_*` that are
absolute counters, distinct from our existing `context_by_mtype`
metrics in that if a gauge is sent for the same context 1,000 times
context_by_mtype will be 1 whereas metric_type_gauge_count will be
1,000.

This also serves as a correctness check on our Regression Detector
experiments: if the metrics mix recorded by DogStatsD is distinct from
the experiment configuration, that's a problem.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>