Got an initial sample of rules_multitool in place.
This lets users bring in prebuild buildifier without having to know what to download.

I have some concerns about rules_multitool, so we'll probably do our
own fork of it to overcome those, but this is good enough for now.

Next step: Add the tool to read the releases for a package and generate the lockfiles.