Support UST tags for Process entities
Support the Unified Service Tagging service, env and version tags for
Process enties in the tagger based on the standard UST environment
variables (DD_SERVICE, DD_ENV, DD_VERSION) collected by the service
discovery code.

Note that the DDService field currently collected by the SD code is not
the same as the DD_SERVICE environment variable since it can come from
other sources such as command line parsing, and it also goes through
normalization.  For now, leave that field alone, and add a separate
field with the raw environment variable value for the UST service
environment variable, similar to what we're adding for env and version.