[AGENTRUN-631] Execute Python check signals on a custom stack
Go's dynamic stack creates problems when installing custom signal
handlers, as the stack may be (and will likely be) too small to do
anything useful.

We discovered this when attempting to gather coredumps and noting that
the agent crashed (without dumping core) inside the signal hander.

This PR creates a separate stack to use for the signal handler and
instructs sigaction to use this custom stack.