dyninst/object: clean up concepts and API
Before this change, there were some sprawling and poorly named abstractions.
The ElfFile exposed an MMappingElfFile but was really about DWARF. There was
also no way to get access to the object and its dwarf without a file path and
mmapping. This was unfortunate.

Now we have two clear interfaces: `File` and `Dwarf`.

When it comes to `File`, there are 3 public implementations:

 * `InMemoryElfFile`: everything in RAM, decompresses on-demand to RAM.
 * `MMappingElfFile`: mmaps sections, never decompresses.
 * `ElfFileWithDwarf`: could be mmapping, could be in ram, decompressed
   data also could be mmapped to temporary disk space or in ram, but it
   does expose compressed section data depending on how it was constructed.
   This is the implementor of Dwarf.

This will enable testing that doesn't need disk space.