fakeintake expose option to command line
The fakeintake server can use some options to be configured, expose
these options on the command line to run the fakeintake server.

Add option to select the data retention period.

Add option to choose between "memory" and "sql" store.

In case of "sql" store, add option to choose the path for the sqlite
database.
This file path can only be set using env variable.
Set the value using the env variable already provided.
In order to preserve current usage, if the env variable is set too, the
env variable will be used.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>