CONTINT-4687 - fakeintake extract host-impl tags
Add necessary code in the fakeintake library to parse and deserialize
the host-tags.

This requires some filtering, the route to read the tags contains other
payloads that could match while calling `Unmarshall`.

Use an inner filed allocated with a pointer to check for its presence
and therefore validate if we are looking at the right payload.

Add necessary code to query the host-implementation tags from the command
line.

Once we have a match return all payloads.

We cannot simply import the `Tags` struct from datadog-agent, as this
will add datadog-agent as dependency on the fakeintake which cascades in
importing the whole repo and its dependencies. To prevent that simply
re-create the matching structure with primitive data type.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>