make `golang_deps_diff` job fully clone the repo (#40118)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

The s3 git strategy (used by default in the agent repo) is basically the
fetch strategy with a different distribution mechanism because of
https://github.com/DataDog/datadog-agent/blob/36a5eec75163ff19617f8f4176eca6b8f9e9138e/.gitlab/.pre/clone.yml#L12.

On the other hand the `golang_deps_diff` job tries to do a `git
merge-base` to compute the base it needs to compare the list of go
dependencies.

Now the issue is that the fetch strategy only fetches the branch it's
working on, not the `main` branch. So the git repo the
`golang_deps_diff` job is working on might have an outdated view of the
main branch, and thus compute an outdated merge base (like a few commits
old) polluting the results.

To fix that, this PR overwrite the git strategy on the job to do a full
clone instead.

### Motivation

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->